Split bedrock streaming into separate smaller modules
